Remove application logic - logout, login, require_login - from helper
