+++ app/controllers/application_controller.rb
@@ -1,3 +1,3 @@
 class ApplicationController < ActionController::Base
-  include SessionsHelper
+  include SessionManagementCapable
 end
