+++ app/controllers/sessions_controller.rb
@@ -11,7 +11,7 @@ class SessionsController < ApplicationController
     if user&.authenticate(params[:session][:password])
       token = SecureRandom.urlsafe_base64
       log_in(token)
-      user.update_attribute(:remember_digest, digest(token))
+      user.update_attribute(:remember_digest, Digest::SHA1.hexdigest(token))
       redirect_to root_url
     else
       flash.now[:danger] = 'Invalid credentials.'
