+++ config/application.rb
@@ -15,5 +15,6 @@ module MembersOnly
     # Application configuration can go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded after loading
     # the framework and any gems in your application.
+    config.autoload_paths += %W[#{config.root}/lib]
   end
 end
